add disposition_history files to SurveyFilesManager
